This is a test program for prototyping a multi-threaded pooling system.
The intended use case is for the Advent 7 (Day 7).
In particular each contorl node needs to run its own program and await
for an input before the passing on the message to a subsequent program.
Prior to implementing this control flow, the int-code system needs to be
refactored so that the functions can used to spawn their own specific threads.

TODO: Day #7